refactor: separate styles generation
