+++ scripts/src/tasks/generateDocumentation.ts
